add: ValueError instead of assert & moove FakeParserBuilder
