+++ megaparse/api/utils/exceptions.py
@@ -16,3 +16,11 @@ class HTTPParsingException(HTTPException):
     def __init__(self, file_name, message="Failed to parse the file"):
         message = f"{file_name} : {message}"
         super().__init__(status_code=500, detail=message)
+
+
+class ParsingException(Exception):
+    """Exception raised for errors in the parsing process."""
+
+    def __init__(self, message="An error occurred during parsing"):
+        self.message = message
+        super().__init__(self.message)
