+++ megaparse/core/parser/builder.py
@@ -29,31 +29,3 @@ class ParserBuilder:
             ValueError: If the configuration is invalid.
         """
         return parser_dict[config.method](**config.model_dump())
-
-
-class FakeParserBuilder:
-    def build(self, config: ParserConfig) -> BaseParser:
-        """
-        Build a fake parser based on the given configuration.
-
-        Args:
-            config (ParserDict): The configuration to be used for building the parser.
-
-        Returns:
-            BaseParser: The built fake parser.
-
-        Raises:
-            ValueError: If the configuration is invalid.
-        """
-
-        class FakeParser(BaseParser):
-            async def convert(
-                self,
-                file_path: str | Path | None = None,
-                file: IO[bytes] | None = None,
-                **kwargs,
-            ) -> str:
-                print("Fake parser is converting the file")
-                return "Fake conversion result"
-
-        return FakeParser()
